fetcher and draft parser
Doesn't do the hard one yet.

fetchitemlist will need modification to put stuff into 
undata - althought that should be quite simple. it creates
its directories for what it needs in the following structure
	

	/undata/[area]/[year or code]/[docid]/
with original.pdf being the (english) document we download,
and out.xml being a badly named xml representation of that.






git-svn-id: http://project.knowledgeforge.net/ukparse/svn/trunk@1205 9cc54934-f9f6-0310-8a8d-e8e977684441